backend/requirements.txt
click==8.0.3
Flask==2.0.2
Flask-Cors==5.0.0
Flask-SQLAlchemy==2.5.1
flask-talisman==0.8.1
gunicorn==19.9.0
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
requests
python-dotenv==0.19.2
six==1.16.0
SQLAlchemy==1.4.27
Werkzeug==2.0.2